Componetization of controls
