+++ examples/cube.html
@@ -15,8 +15,9 @@
     <vr-object-template id="sphere" geometry="primitive: sphere" material="color: blue"></vr-object-template>
   </vr-assets>
   <vr-scene>
-    <vr-object position="0 0 20" camera="fov: 45"></vr-object>
-    <vr-object template="cube" position="-10 0 0" rotation="0 45 0"></vr-object>
+    <vr-object position="0 0 20" camera="fov: 45" controls="mouselook: true; locomotion: true"></vr-object>
+    <!-- specialization of a particular component with an applied template -->
+    <vr-object template="cube" position="-10 0 0" rotation="0 45 0" material="color: yellow"></vr-object>
     <vr-object template="cube" position="0 0 0" rotation="0 45 0"></vr-object>
     <vr-object template="cube" position="10 0 0" rotation="0 45 0"></vr-object>
   </vr-scene>
