+++ src/components/components.js
@@ -1,3 +1,4 @@
 module.exports.camera = require('./camera');
+module.exports.controls = require('./controls');
 module.exports.geometry = require('./geometry');
 module.exports.material = require('./material');
