+++ src/vr-markup.js
@@ -15,12 +15,8 @@ require('./core/vr-scene');
 
 require('./vr-animation');
 require('./vr-behavior');
-require('./vr-controls');
 require('./vr-cursor');
 require('./vr-fog');
-require('./vr-geometry');
-require('./vr-material');
-require('./vr-mesh');
 
 module.exports = {
   THREE: THREE,
